refactor: removed unused dataset GraphQL query
