+++ knowledge-graph/README.md
@@ -263,167 +263,6 @@ Response body example:
 }
 ```
 
-* Data-sets
-
-Query example:
-```
-{
-  "query": "{ 
-    datasets(projectPath: \"namespace/project\") {
-      identifier!
-      name!
-      description
-      published {
-        datePublished
-        creator {
-          name!
-          email
-        }
-      }
-      hasPart {
-        name!
-        atLocation!
-      }
-      isPartOf {
-        path!
-        name!
-        created! {
-          dateCreated! 
-          agent { 
-            email!
-            name!
-          }
-        }
-      }
-    } 
-  }"
-}
-
-```
-Response body example:
-```
-{
-  "data": {
-    "datasets": [
-      {
-        "identifier": "e1fc7b62-e021-434b-9264-dda336bddd4f",
-        "name": "dataset name",
-        "description": "Data-set long description",
-        "published": {
-          "datePublished": "2019-07-30",
-          "creator": []
-        },
-        "hasPart": [],
-        "isPartOf": [{
-          "path": "namespace/project",
-          "name": "project name",
-          "created" : {
-            "dateCreated" : "1970-05-12T06:06:41.448Z",
-            "agent" : {
-              "email" : "n@ulQdsXl",
-              "name" : "v imzn"
-            }
-          }
-        }]
-      },
-      {
-        "identifier": "5b7a1394-93b5-4e75-932d-e041cf46349d",
-        "name": "xr",
-        "description": null,
-        "created": {
-          "dateCreated": "1991-09-05T10:38:29.457Z",
-          "agent": {
-            "email": "user2@host",
-            "name": "user 2"
-          }
-        },
-        "published": {
-          "datePublished": null,
-          "creator": [{
-            "name": "author1 name",
-            "email": null
-          }]
-        },
-        "hasPart": [{
-          "name": "file1",
-          "atLocation"": "data/dataset-name/file1"
-        }],
-        "isPartOf": [{
-          "path": "namespace/project",
-          "name": "project name",
-          "created" : {
-            "dateCreated" : "1970-05-12T06:06:41.448Z",
-            "agent" : {
-              "email" : "n@ulQdsXl",
-              "name" : "v imzn"
-            }
-          }
-        }]
-      },
-      {
-        "identifier": "b1aa58af-a488-4bae-97b4-d6d349f98412",
-        "name": "chOorWhraw",
-        "description": null,
-        "created": {
-          "dateCreated": "2001-09-05T10:38:29.457Z",
-          "agent": {
-            "email": "user3@host",
-            "name": "user 3"
-          }
-        },
-        "published": {
-          "datePublished": "2001-09-05T10:38:29.457Z",
-          "creator": [{
-            "name": "author1 name",
-            "email": "author1@mail.org"
-          }, {
-            "name": "author2 name",
-            "email": "author2@mail.org"
-          }]
-        },
-        "hasPart": [{
-          "name": "file1",  
-          "atLocation"": "data/chOorWhraw-name/file1"
-        }, {
-          "name": "file2"  
-          "atLocation"": "data/chOorWhraw-name/file2"
-        }],
-        "isPartOf": [{
-          "path": "namespace/project1",
-          "name": "project1 name",
-          "created" : {
-            "dateCreated" : "1970-10-12T06:06:41.448Z",
-            "agent" : {
-              "email" : "n@ulQdsXl",
-              "name" : "v imzn"
-            }
-          }
-        }, {
-          "path": "namespace/project2",
-          "name": "project2 name",
-          "created" : {
-            "dateCreated" : "1970-05-12T06:06:41.448Z",
-            "agent" : {
-              "email" : "n@ulQdsXl",
-              "name" : "v imzn"
-            }
-          }
-        }]
-      }
-    ]
-  }
-}
-```
-
-In case there's no data found for a given query, the response `json` will contain a property with the queried resource name and a `null` value. For instance if a user performs a lineage query and there's no lineage found, the response will look like below:
-```
-{
-  "data": {
-    "datasets": null
-  }
-}
-```
-
 #### GET /knowledge-graph/projects/:namespace/:name
 
 Finds details of the project with the given `namespace/name`.
