--- knowledge-graph/src/main/scala/ch/datascience/knowledgegraph/datasets/graphql/QueryFields.scala
@@ -1,45 +1,0 @@
-/*
- * Copyright 2020 Swiss Data Science Center (SDSC)
- * A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
- * Eidgenössische Technische Hochschule Zürich (ETHZ).
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ch.datascience.knowledgegraph.datasets.graphql
-
-import cats.effect.IO
-import ch.datascience.knowledgegraph.graphql.CommonQueryFields._
-import ch.datascience.knowledgegraph.graphql.QueryContext
-import sangria.schema._
-
-import scala.language.higherKinds
-
-object QueryFields {
-
-  import modelSchema._
-
-  def apply(): List[Field[QueryContext[IO], Unit]] =
-    fields[QueryContext[IO], Unit](
-      Field(
-        name        = "datasets",
-        fieldType   = ListType(datasetType),
-        description = Some("Returns datasets defined in the project"),
-        arguments   = List(projectPathArgument),
-        resolve = context =>
-          context.ctx.datasetsFinder
-            .findDatasets(context.args arg projectPathArgument)
-            .unsafeToFuture()
-      )
-    )
-}
