+++ knowledge-graph/src/main/scala/ch/datascience/knowledgegraph/graphql/QueryContext.scala
@@ -19,15 +19,13 @@
 package ch.datascience.knowledgegraph.graphql
 
 import cats.effect.{ContextShift, IO, Timer}
-import ch.datascience.knowledgegraph.datasets.graphql.{IOProjectDatasetsFinder, ProjectDatasetsFinder}
 import ch.datascience.knowledgegraph.lineage.{IOLineageFinder, LineageFinder}
 
 import scala.concurrent.ExecutionContext
 import scala.language.higherKinds
 
 class QueryContext[Interpretation[_]](
-    val lineageFinder:  LineageFinder[Interpretation],
-    val datasetsFinder: ProjectDatasetsFinder[Interpretation]
+    val lineageFinder: LineageFinder[Interpretation]
 )
 
 object IOQueryContext {
@@ -35,7 +33,6 @@ object IOQueryContext {
               contextShift:              ContextShift[IO],
               timer:                     Timer[IO]): IO[QueryContext[IO]] =
     for {
-      lineageFinder  <- IOLineageFinder()
-      datasetsFinder <- IOProjectDatasetsFinder()
-    } yield new QueryContext[IO](lineageFinder, datasetsFinder)
+      lineageFinder <- IOLineageFinder()
+    } yield new QueryContext[IO](lineageFinder)
 }
