+++ knowledge-graph/src/test/scala/ch/datascience/knowledgegraph/graphql/QuerySchemaSpec.scala
@@ -19,18 +19,14 @@
 package ch.datascience.knowledgegraph.graphql
 
 import cats.effect.IO
-import ch.datascience.generators.Generators._
 import ch.datascience.graph.model.events.CommitId
 import ch.datascience.graph.model.projects.{FilePath, ProjectPath}
-import ch.datascience.knowledgegraph.datasets.DatasetsGenerators
-import ch.datascience.knowledgegraph.datasets.graphql.ProjectDatasetsFinder
-import ch.datascience.knowledgegraph.datasets.model._
+import ch.datascience.knowledgegraph.lineage
 import ch.datascience.knowledgegraph.lineage.LineageFinder
 import ch.datascience.knowledgegraph.lineage.model.Node.{SourceNode, TargetNode}
 import ch.datascience.knowledgegraph.lineage.model._
-import ch.datascience.knowledgegraph.{datasets, lineage}
+import io.circe.Json
 import io.circe.literal._
-import io.circe.{Encoder, Json}
 import org.scalamock.scalatest.MockFactory
 import org.scalatest.Matchers._
 import org.scalatest.WordSpec
@@ -73,39 +69,17 @@ class QuerySchemaSpec
 
       execute(query) shouldBe json(lineage)
     }
-
-    "allow to search for project's datasets" in new DatasetsTestCase {
-      forAll(datasetsLists) { datasets =>
-        val query = graphql"""
-        {
-          datasets(projectPath: "namespace/project") {
-            identifier
-            name
-            description
-            published { datePublished creator { email name } }
-            hasPart { name atLocation }
-            isPartOf { path name created { dateCreated agent { email name } } }
-          }
-        }"""
-
-        givenFindDatasets(ProjectPath("namespace/project"))
-          .returning(IO.pure(datasets))
-
-        execute(query) shouldBe json(datasets)
-      }
-    }
   }
 
   private trait TestCase {
-    val lineageFinder  = mock[LineageFinder[IO]]
-    val datasetsFinder = mock[ProjectDatasetsFinder[IO]]
+    val lineageFinder = mock[LineageFinder[IO]]
 
     def execute(query: Document): Json =
       Executor
         .execute(
-          QuerySchema[IO](lineage.graphql.QueryFields(), datasets.graphql.QueryFields()),
+          QuerySchema[IO](lineage.graphql.QueryFields()),
           query,
-          new QueryContext[IO](lineageFinder, datasetsFinder)
+          new QueryContext[IO](lineageFinder)
         )
         .futureValue
   }
@@ -149,68 +123,5 @@ class QuerySchemaSpec
           "source" : ${edge.source.id.value},
           "target" : ${edge.target.id.value}
         }"""
-  }
-
-  private trait DatasetsTestCase extends TestCase {
-
-    def givenFindDatasets(projectPath: ProjectPath) = new {
-      def returning(result: IO[List[Dataset]]) =
-        (datasetsFinder
-          .findDatasets(_: ProjectPath))
-          .expects(projectPath)
-          .returning(result)
-    }
-
-    lazy val datasetsLists = nonEmptyList(DatasetsGenerators.datasets).map(_.toList)
-
-    def json(datasets: List[Dataset]) = json"""
-        {
-          "data" : {
-            "datasets" : ${Json.arr(datasets.map(toJson): _*)}
-          }
-        }"""
-
-    // format: off
-    private def toJson(dataset: Dataset): Json = json"""
-      {
-        "identifier": ${dataset.id.value},
-        "name": ${dataset.name.value},
-        "description": ${dataset.maybeDescription.map(_.value).map(Json.fromString).getOrElse(Json.Null)},
-        "published": {
-          "datePublished": ${dataset.published.maybeDate.map(_.toString).map(Json.fromString).getOrElse(Json.Null)},
-          "creator": ${dataset.published.creators.toList}
-        },
-        "hasPart": ${dataset.parts},
-        "isPartOf": ${dataset.projects}
-      }"""
-    // format: on
-
-    private implicit lazy val creatorEncoder: Encoder[DatasetCreator] = Encoder.instance[DatasetCreator] { creator =>
-      json"""{
-        "email": ${creator.maybeEmail.map(_.toString).map(Json.fromString).getOrElse(Json.Null)},
-        "name": ${creator.name.value}
-      }"""
-    }
-
-    private implicit lazy val partEncoder: Encoder[DatasetPart] = Encoder.instance[DatasetPart] { part =>
-      json"""{
-        "name": ${part.name.value},
-        "atLocation": ${part.atLocation.value}
-      }"""
-    }
-
-    private implicit lazy val projectEncoder: Encoder[DatasetProject] = Encoder.instance[DatasetProject] { project =>
-      json"""{
-        "path": ${project.path.value},
-        "name": ${project.name.value},
-        "created": {
-          "dateCreated": ${project.created.date.value},
-          "agent": {
-            "email": ${project.created.agent.email.value},
-            "name": ${project.created.agent.name.value}
-          }
-        }
-      }"""
-    }
   }
 }
